feat(frontend, backend): implement association management and enhance canvas data handling
