+++ app/frontend/src/hooks/useCanvasMouseEvents.ts
@@ -2,9 +2,15 @@ import React, { RefObject } from "react";
 import { ToPoint } from "../utils/wailsBridge";
 import { SelectComponent } from "../../wailsjs/go/umlproject/UMLProject";
 
+interface ExtraHandlers {
+    onCanvasClick?: (point: {x: number, y: number}) => void;
+    isAddingAssociation?: boolean;
+}
+
 export function useCanvasMouseEvents(
     canvasRef: RefObject<HTMLCanvasElement>,
-    onSelect: () => void
+    onSelect: () => void,
+    extraHandlers?: ExtraHandlers
 ) {
     const handleMouseDown = (event: React.MouseEvent<HTMLCanvasElement>) => {
         const canvas = canvasRef.current;
@@ -19,13 +25,17 @@ export function useCanvasMouseEvents(
         const x = Math.round((event.clientX - rect.left) * scaleX);
         const y = Math.round((event.clientY - rect.top) * scaleY);
 
-        SelectComponent(ToPoint(x, y))
-            .then(() => {
-                onSelect();
-            })
-            .catch((error) => {
-                console.error("Error selecting component:", error);
-            });
+        if (extraHandlers?.isAddingAssociation && extraHandlers?.onCanvasClick) {
+            extraHandlers.onCanvasClick({x, y});
+        } else {
+            SelectComponent(ToPoint(x, y))
+                .then(() => {
+                    onSelect();
+                })
+                .catch((error) => {
+                    console.error("Error selecting component:", error);
+                });
+        }
     };
 
     const handleMouseMove = (event: React.MouseEvent<HTMLCanvasElement>) => {
